fix cast speed to uint

+++ libraries/HoloMove/HoloMove.cpp
@@ -43,9 +43,9 @@ void HoloMove::move(float direction, float speed, float rotation, float *a, floa
         *d = 0;
     }
 
-    _rb->ForwardBackwardM1(0x80, map(*a, -255, 255, 0, 127));
-    _rb->ForwardBackwardM1(0x81, map(*c, -255, 255, 0, 127));
-    _rb->ForwardBackwardM2(0x81, map(*d, -255, 255, 0, 127));
-    _rb->ForwardBackwardM2(0x80, map(*b, -255, 255, 0, 127));
+    _rb->ForwardBackwardM1(0x80, (uint8_t) map(*a, -255, 255, 0, 127));
+    _rb->ForwardBackwardM1(0x81, (uint8_t) map(*c, -255, 255, 0, 127));
+    _rb->ForwardBackwardM2(0x81, (uint8_t) map(*d, -255, 255, 0, 127));
+    _rb->ForwardBackwardM2(0x80, (uint8_t) map(*b, -255, 255, 0, 127));
 
 }
